Category-wide warning filters now use u8 instead of Category type

+++ external-crates/move/move-compiler/src/expansion/translate.rs
@@ -739,7 +739,7 @@ fn warning_filter(
                 let msg = format!(
                     "Expected list of warnings, e.g. '{}({})'",
                     DiagnosticAttribute::ALLOW,
-                    WarningFilter::Category(Category::UnusedItem, Some(FILTER_UNUSED))
+                    WarningFilter::Category(Category::UnusedItem as u8, Some(FILTER_UNUSED))
                         .to_str()
                         .unwrap(),
                 );
